requirements.txt
opencv-python~=4.8.1.78
opencv-contrib-python
Flask~=3.0.0
numpy~=1.26.2
Flask_SocketIO
mediapipe~=0.10.8
paho-mqtt~=1.6.1
firebase_admin
Flask_Principal
Flask_Login
scikit-learn~=1.3.2
matplotlib~=3.8.2
Pillow~=10.1.0